+++ src/Html2OpenXml/Expressions/Table/TableExpression.cs
@@ -9,10 +9,7 @@
  * IMPLIED WARRANTIES OF MERCHANTABILITY AND/OR FITNESS FOR A
  * PARTICULAR PURPOSE.
  */
-using System;
-using System.Collections.Generic;
 using System.Globalization;
-using System.Linq;
 using AngleSharp.Html.Dom;
 using DocumentFormat.OpenXml;
 using DocumentFormat.OpenXml.Wordprocessing;
@@ -169,10 +166,10 @@ sealed class TableExpression(IHtmlTableElement node) : PhrasingElementExpression
 
         styleAttributes = tableNode.GetStyles();
         var width = styleAttributes.GetUnit("width", UnitMetric.Pixel);
-        if (!width.IsValid) width = Unit.Parse(tableNode.GetAttribute("width"), UnitMetric.Pixel);
+        if (!width.IsValid) width = Unit.Parse(tableNode.GetAttribute("width").AsSpan(), UnitMetric.Pixel);
         if (!width.IsValid) width = new Unit(UnitMetric.Percent, 100);
 
-        switch (width.Type)
+        switch (width.Metric)
         {
             case UnitMetric.Percent:
                 tableProperties.TableWidth = new TableWidth
@@ -286,9 +283,9 @@ sealed class TableExpression(IHtmlTableElement node) : PhrasingElementExpression
         if (!align.HasValue)
         {
             var margin = styleAttributes.GetMargin("margin");
-            if (margin.Left.Type == UnitMetric.Auto)
+            if (margin.Left.Metric == UnitMetric.Auto)
             {
-                if (margin.Right.Type == UnitMetric.Auto)
+                if (margin.Right.Metric == UnitMetric.Auto)
                     align = JustificationValues.Center;
                 else
                     align = JustificationValues.Right;