+++ src/Html2OpenXml/IO/Resource.cs
@@ -9,9 +9,6 @@
  * IMPLIED WARRANTIES OF MERCHANTABILITY AND/OR FITNESS FOR A
  * PARTICULAR PURPOSE.
  */
-using System;
-using System.Collections.Generic;
-using System.IO;
 using System.Net;
 
 namespace HtmlToOpenXml.IO;