+++ src/Html2OpenXml/IO/SequentialBinaryReader.cs
@@ -12,7 +12,6 @@
  * Inspiration from Metadata Extractor (Drew Noakes):
  * https://github.com/drewnoakes/metadata-extractor-dotnet
  */
-using System.IO;
 
 namespace HtmlToOpenXml.IO;