+++ test/HtmlToOpenXml.Tests/Primitives/MarginTests.cs
@@ -12,12 +12,13 @@ namespace HtmlToOpenXml.Tests.Primitives
         [TestCase("25px 50px 75px", 25, 50, 75, 50)]
         [TestCase("25px 50px", 25, 50, 25, 50)]
         [TestCase("25px", 25, 25, 25, 25)]
+        [TestCase("25px   75px", 25, 75, 25, 75)]
         public void ParseHtmlString_ShouldSucceed (string html, int top, int right, int bottom, int left)
         {
-            var margin = Margin.Parse(html);
+            var margin = Margin.Parse(html.AsSpan());
 
             Assert.Multiple(() => {
-                Assert.That(margin.IsValid, Is.EqualTo(true));
+                Assert.That(margin.IsValid, Is.True);
                 Assert.That(margin.Top.ValueInPx, Is.EqualTo(top));
                 Assert.That(margin.Right.ValueInPx, Is.EqualTo(right));
                 Assert.That(margin.Bottom.ValueInPx, Is.EqualTo(bottom));
@@ -28,25 +29,25 @@ namespace HtmlToOpenXml.Tests.Primitives
         [Test]
         public void ParseWithFloat_ShouldSucceed ()
         {
-            var margin = Margin.Parse("0 50% 9.5pt .00001pt");
+            var margin = Margin.Parse("0 50% 9.5pt .00001pt".AsSpan());
 
             Assert.Multiple(() => {
-                Assert.That(margin.IsValid, Is.EqualTo(true));
+                Assert.That(margin.IsValid, Is.True);
 
-                Assert.That(margin.Top.Value, Is.EqualTo(0));
-                Assert.That(margin.Top.Type, Is.EqualTo(UnitMetric.Pixel));
+                Assert.That(margin.Top.Value, Is.Zero);
+                Assert.That(margin.Top.Metric, Is.EqualTo(UnitMetric.Pixel));
 
                 Assert.That(margin.Right.Value, Is.EqualTo(50));
-                Assert.That(margin.Right.Type, Is.EqualTo(UnitMetric.Percent));
+                Assert.That(margin.Right.Metric, Is.EqualTo(UnitMetric.Percent));
 
                 Assert.That(margin.Bottom.Value, Is.EqualTo(9.5));
-                Assert.That(margin.Bottom.Type, Is.EqualTo(UnitMetric.Point));
+                Assert.That(margin.Bottom.Metric, Is.EqualTo(UnitMetric.Point));
                 Assert.That(margin.Bottom.ValueInPoint, Is.EqualTo(9.5));
                 //size are half-point font size (OpenXml relies mostly on long value, not on float)
                 Assert.That(Math.Round(margin.Bottom.ValueInPoint * 2).ToString(), Is.EqualTo("19"));
 
                 Assert.That(margin.Left.Value, Is.EqualTo(.00001));
-                Assert.That(margin.Left.Type, Is.EqualTo(UnitMetric.Point));
+                Assert.That(margin.Left.Metric, Is.EqualTo(UnitMetric.Point));
                 // but due to conversion: 0 (OpenXml relies mostly on long value, not on float)
                 Assert.That(Math.Round(margin.Left.ValueInPoint * 2).ToString(), Is.EqualTo("0"));
             });
@@ -55,19 +56,19 @@ namespace HtmlToOpenXml.Tests.Primitives
         [Test]
         public void ParseWithAuto_ShouldSucceed ()
         {
-            var margin = Margin.Parse("0 auto");
+            var margin = Margin.Parse("0 auto".AsSpan());
 
             Assert.Multiple(() => {
-                Assert.That(margin.IsValid, Is.EqualTo(true));
+                Assert.That(margin.IsValid, Is.True);
 
-                Assert.That(margin.Top.Value, Is.EqualTo(0));
-                Assert.That(margin.Top.Type, Is.EqualTo(UnitMetric.Pixel));
+                Assert.That(margin.Top.Value, Is.Zero);
+                Assert.That(margin.Top.Metric, Is.EqualTo(UnitMetric.Pixel));
 
-                Assert.That(margin.Bottom.Value, Is.EqualTo(0));
-                Assert.That(margin.Bottom.Type, Is.EqualTo(UnitMetric.Pixel));
+                Assert.That(margin.Bottom.Value, Is.Zero);
+                Assert.That(margin.Bottom.Metric, Is.EqualTo(UnitMetric.Pixel));
 
-                Assert.That(margin.Left.Type, Is.EqualTo(UnitMetric.Auto));
-                Assert.That(margin.Right.Type, Is.EqualTo(UnitMetric.Auto));
+                Assert.That(margin.Left.Metric, Is.EqualTo(UnitMetric.Auto));
+                Assert.That(margin.Right.Metric, Is.EqualTo(UnitMetric.Auto));
             });
         }
     }